feat(fix-harness): real step logic in Behave session steps
Move initiator creation from before_scenario into the when step.
given configures CompIDs on context; when builds the QuickFIX cfg
and starts the initiator; then asserts the expected state.
environment.py reduced to lifecycle hooks only.
